Remove ExSwift
Previously, because of the way ExSwift had to be included in the
framework, its extensions leaked out as public modifications to
existing APIs.  Other projects that also used ExSwift would then have
conflicts and ambiguous method invocations.  This change removes
ExSwift from the project, adding in only the few extensions that were
required and limiting their visibility to the framework itself.